fix: pass highlight values from frame to quad

+++ pudu_ui/styles/frames.py
@@ -36,7 +36,7 @@ class FrameStyle:
     radius_bottom_left: float = 0
     radius_bottom_right: float = 0
     gradient_direction: GradientDirection = GradientDirection.VERTICAL
-    highlight_width: float = 3
+    highlight_width: int = 3
     highlight_color: Color = field(default_factory=highlight_color)
 
     def set_solid_color(self, color: Color):
